+++ src/lib/foodSpotImages.ts
@@ -1,0 +1,9 @@
+// 精选名店照片：scripts/fetch-foodspot-images.mts 从 Hot Pepper 按店名+坐标匹配（真实 URL，无则缺省）。
+export const FOOD_SPOT_IMAGES: Record<string, string> = {
+  "ryugin": "https://upload.wikimedia.org/wikipedia/commons/thumb/8/84/RyuGin.jpg/1280px-RyuGin.jpg",
+  "koju": "https://imgfp.hotp.jp/IMGH/61/28/P012106128/P012106128_238.jpg",
+  "sazenka": "https://sazenka.com/wp-content/themes/sazenka/img/ogp.jpg",
+  "narisawa": "https://user-images.strikinglycdn.com/res/hrscywv4p/image/upload/c_limit,fl_lossy,h_630,w_1200,f_auto,q_auto/1334583/%E3%83%8F%E3%83%81_ssuiqm.jpg",
+  "bluebottle": "https://cdn.shopify.com/s/files/1/0288/0549/6880/files/OGP_b7656059-c938-4a31-b904-996595535529.jpg?v=1631679476",
+  "higashiya": "https://www.higashiya.com/assets/img/common/meta/facebook-ogp-image.png"
+};